Add files via upload
For this project, we worked to understand the results of an A/B test run by an e-commerce website. The company had developed a new web page in order to try and increase the number of users who "convert," meaning the number of users who decide to pay for the company's product. Our goal was to help the company understand if they should implement this new page, keep the old page, or perhaps run the experiment longer to make their decision.